+++ ruoyi-order/src/main/java/com/ruoyi/order/service/IRfOrderService.java
@@ -5,23 +5,23 @@ import com.ruoyi.order.domain.RfOrder;
 
 /**
  * 采购订单Service接口
- * 
+ *
  * @author pg
- * @date 2024-01-13
+ * @date 2024-01-17
  */
-public interface IRfOrderService 
+public interface IRfOrderService
 {
     /**
      * 查询采购订单
-     * 
+     *
      * @param id 采购订单主键
      * @return 采购订单
      */
-    public RfOrder selectRfOrderById(Long id);
+    public RfOrder selectRfOrderById(Integer id);
 
     /**
      * 查询采购订单列表
-     * 
+     *
      * @param rfOrder 采购订单
      * @return 采购订单集合
      */
@@ -29,7 +29,7 @@ public interface IRfOrderService
 
     /**
      * 新增采购订单
-     * 
+     *
      * @param rfOrder 采购订单
      * @return 结果
      */
@@ -37,7 +37,7 @@ public interface IRfOrderService
 
     /**
      * 修改采购订单
-     * 
+     *
      * @param rfOrder 采购订单
      * @return 结果
      */
@@ -45,7 +45,7 @@ public interface IRfOrderService
 
     /**
      * 批量删除采购订单
-     * 
+     *
      * @param ids 需要删除的采购订单主键集合
      * @return 结果
      */
@@ -53,9 +53,9 @@ public interface IRfOrderService
 
     /**
      * 删除采购订单信息
-     * 
+     *
      * @param id 采购订单主键
      * @return 结果
      */
-    public int deleteRfOrderById(Long id);
+    public int deleteRfOrderById(Integer id);
 }
